Heed a compiler warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@59988 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/netcache/test/test_nc_stress.cpp
@@ -510,9 +510,9 @@ int CTestNetCacheStressApp::Run(void)
         threads = (unsigned) args["threads"].AsInteger();
     }
 
-    int timeout = 2000;
+    unsigned timeout = 2000;
     if (args["timeout"]) {
-        timeout = args["timeout"].AsInteger();
+        timeout = (unsigned) args["timeout"].AsInteger();
     }
